+++ db/schema.rb
@@ -80,8 +80,8 @@ ActiveRecord::Schema.define(version: 2021_05_29_171735) do
     t.string "title", null: false
     t.string "short_description", null: false
     t.text "description", null: false
-    t.decimal "price", null: false
-    t.decimal "old_price"
+    t.integer "price", null: false
+    t.integer "old_price"
     t.string "dimentions"
     t.integer "difficulty", null: false
     t.string "materials"
